PythonSDK/NxBCI/requirements.txt
async-timeout==4.0.3
bleak==0.22.3
bleak-winrt==1.2.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==1.24.4
packaging==25.0
paho-mqtt==2.1.0
pillow==11.3.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.13.2
zipp==3.23.0